+++ Packagers/package-thunderbird.ps1
@@ -183,16 +183,16 @@ function Invoke-StageThunderbird {
     }
 
     # --- Generate content wrappers (custom install args) ---
-    $installPs1 = @(
-        "`$msiPath = Join-Path `$PSScriptRoot '$MsiFileName'"
-        "`$proc = Start-Process msiexec.exe -ArgumentList @('/i', `"``\`"`$msiPath``\`"`", '/quiet', '/norestart', 'INSTALL_MAINTENANCE_SERVICE=false', 'TASKBAR_SHORTCUT=false', 'DESKTOP_SHORTCUT=false') -Wait -PassThru -NoNewWindow"
-        "exit `$proc.ExitCode"
+    $installPs1 = (
+        ('$msiPath = Join-Path $PSScriptRoot ''{0}''' -f $MsiFileName),
+        '$proc = Start-Process msiexec.exe -ArgumentList @(''/i'', "`"$msiPath`"", ''/quiet'', ''/norestart'', ''INSTALL_MAINTENANCE_SERVICE=false'', ''TASKBAR_SHORTCUT=false'', ''DESKTOP_SHORTCUT=false'') -Wait -PassThru -NoNewWindow',
+        'exit $proc.ExitCode'
     ) -join "`r`n"
 
-    $uninstallPs1 = @(
-        "`$msiPath = Join-Path `$PSScriptRoot '$MsiFileName'"
-        "`$proc = Start-Process msiexec.exe -ArgumentList @('/x', `"``\`"`$msiPath``\`"`", '/quiet', '/norestart') -Wait -PassThru -NoNewWindow"
-        "exit `$proc.ExitCode"
+    $uninstallPs1 = (
+        ('$msiPath = Join-Path $PSScriptRoot ''{0}''' -f $MsiFileName),
+        '$proc = Start-Process msiexec.exe -ArgumentList @(''/x'', "`"$msiPath`"", ''/quiet'', ''/norestart'') -Wait -PassThru -NoNewWindow',
+        'exit $proc.ExitCode'
     ) -join "`r`n"
 
     Write-ContentWrappers -OutputPath $localContentPath `